// types/record.type.ts
export interface Record {
  inetnum: string // 100%
  netname: string // 100%
  descr: string // 100%
  'admin-c': string // 100%
  'tech-c': string // 100%
  status: string // 100%
  'mnt-by': string // 100%
  changed: string // 100%
  source: string // 100%
  country?: string // 99%
  notify?: string // 35%
  'mnt-lower'?: string // 14%
  remarks?: string // 14%
  org?: string // 6%
}